Replace size_t with ee_size_t in barebones/ee_printf.c
Does not change functionality for barebones/ee_printf.c. However, if this file is used as a basis for other ports' printf implementation, compilation might fail, since size_t may not be defined.

This commit modifies size_t to ee_size_t, a type whcih the benchmark requires the user to implement, with equal semantics.