Print total_time as unsigned long long
CoreMark requires calculations to be performed for at least
10 seconds. So for processors with frequency > 429.5 MHz
the total time in processor's ticks will exceed the size
of 32-bit variable.

Not for all platforms sizeof(long unsigned) = 8 bytes. The
C language standard guarantees the minimum size for types.
For unsigned long the minimum size is 4 bytes so the output
of the total_time may be incorrect.

Type long long unsigned should be used for which 8 byte
size is guaranteed by C language standard.